pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.3",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-nmap>=0.7.1",
    "paramiko>=3.5.1",
    "requests>=2.32.3",
    "stem>=1.8.2",
    "pysocks>=1.7.1",
    "jinja2>=3.1.6",
]
